modules/engine.py
import numpy as np
from Bio.PDB import PDBParser, PDBIO, PDBList
from anarci import anarci
from rdkit import Chem
from rdkit.Chem import AllChem
# OpenMM / Vina 関連のインポートは環境に合わせて調整

class GlycoEngine:
    def __init__(self, template_pdb):
        self.template_pdb = template_pdb
        self.parser = PDBParser(QUIET=True)

    def extract_cdrs(self, chain_id):
        # ANARCIを用いたCDR抽出ロジック（前述のコードを統合）
        pass

    def graft_and_optimize(self, mutations):
        """CDR置換とOpenMMによる最適化"""
        # PDBFixerでのIndels対応 + OpenMMでのエネルギー最小化
        # E_total = E_vdw + E_elec + E_hbond + E_solvation
        pass

    def run_docking(self, smiles, ensemble_size=20):
        """アンサンブルドッキングと水分子考慮スコアリング"""
        # Meeko, Vina を用いたシミュレーション
        # 保存水の保持ロジックをここに実装
        return {"affinity": -8.5, "h_bonds": 5}