requirements.txt
streamlit
biopython
requests
numpy
pandas
matplotlib
plotly
fpdf
rdkit
meeko
anarci @ git+https://github.com/oxpig/ANARCI.git
# openmm, pdbfixer, anarci は conda でのインストールを推奨